packages/metadata/python/pyproject.toml
[project]
name = "domain-explorer-metadata"
version = "0.1.0"
description = "Pydantic v2 schemas + YAML loader for the Domain Explorer metadata registry."
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Vinayak Pai", email = "kvinayakpai@gmail.com" }]
dependencies = [
  "pydantic>=2.7",
  "PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["metadata"]
